docs: fix residue mode illustration rendering (#281)
The existing lil ASCII table in the ResidueMode docstring wouldn't render properly in Sphinx nor in markdown-style blocks (eg in VSCode popups). This adds an RST table and retains the ASCII table but puts it in a code block, and then strips the ASCII table out when constructing the Sphinx autodoc page.