fix: don't quote an already quoted url field (#433)
At present, CST doesn't connect to UTA databases correctly if the
password has special characters.

The database URL includes a password field, which is already embedded
within a valid URL. The value was url quoted when the URL was generated,
so it doesn't need to be quoted a second time when the URL is consumed
by the UTA database code.

I tested this using a password `local@$%dev` and it attempted to
authenticate with the password `local%40%24%25dev`. After this PR, it
correctly uses the single-unquoted value.